+++ cache/in-memory/src/updates.rs
@@ -919,6 +919,8 @@ mod tests {
             pinned: false,
             reactions: Vec::new(),
             reference: None,
+            stickers: Vec::new(),
+            referenced_message: None,
             timestamp: String::new(),
             tts: false,
             webhook_id: None,