+++ http/src/request/channel/update_channel_permission_configured.rs
@@ -1,11 +1,16 @@
 use crate::request::prelude::*;
-use twilight_model::{guild::Permissions, id::ChannelId};
+use twilight_model::{
+    channel::permission_overwrite::{PermissionOverwriteTargetType, PermissionOverwriteType},
+    guild::Permissions,
+    id::ChannelId,
+};
 
 #[derive(Serialize)]
 struct UpdateChannelPermissionConfiguredFields {
     allow: Permissions,
     deny: Permissions,
-    kind: String,
+    #[serde(rename = "type")]
+    kind: PermissionOverwriteTargetType,
 }
 
 /// Created when either `member` or `role` is called on a `DeleteChannelPermission` struct.
@@ -24,15 +29,23 @@ impl<'a> UpdateChannelPermissionConfigured<'a> {
         channel_id: ChannelId,
         allow: Permissions,
         deny: Permissions,
-        kind: impl Into<String>,
-        target_id: u64,
+        target: &PermissionOverwriteType,
     ) -> Self {
+        let (name, target_id) = match target {
+            PermissionOverwriteType::Member(user_id) => {
+                (PermissionOverwriteTargetType::Member, user_id.0)
+            }
+            PermissionOverwriteType::Role(role_id) => {
+                (PermissionOverwriteTargetType::Role, role_id.0)
+            }
+        };
+
         Self {
             channel_id,
             fields: UpdateChannelPermissionConfiguredFields {
                 allow,
                 deny,
-                kind: kind.into(),
+                kind: name,
             },
             fut: None,
             http,
@@ -41,8 +54,8 @@ impl<'a> UpdateChannelPermissionConfigured<'a> {
         }
     }
 
-    fn start(&mut self) -> Result<()> {
-        let request = if let Some(reason) = &self.reason {
+    fn request(&self) -> Result<Request> {
+        Ok(if let Some(reason) = &self.reason {
             let headers = audit_header(&reason)?;
             Request::from((
                 crate::json_to_vec(&self.fields)?,
@@ -60,7 +73,11 @@ impl<'a> UpdateChannelPermissionConfigured<'a> {
                     target_id: self.target_id,
                 },
             ))
-        };
+        })
+    }
+
+    fn start(&mut self) -> Result<()> {
+        let request = self.request()?;
 
         self.fut.replace(Box::pin(self.http.verify(request)));
 
@@ -78,3 +95,42 @@ impl<'a> AuditLogReason for UpdateChannelPermissionConfigured<'a> {
 }
 
 poll_req!(UpdateChannelPermissionConfigured<'_>, ());
+
+#[cfg(test)]
+mod tests {
+    use super::{UpdateChannelPermissionConfigured, UpdateChannelPermissionConfiguredFields};
+    use crate::{request::Request, routing::Route, Client};
+    use twilight_model::{
+        channel::permission_overwrite::{PermissionOverwriteTargetType, PermissionOverwriteType},
+        guild::Permissions,
+        id::{ChannelId, UserId},
+    };
+
+    #[test]
+    fn test_request() {
+        let client = Client::new("foo");
+        let builder = UpdateChannelPermissionConfigured::new(
+            &client,
+            ChannelId(1),
+            Permissions::empty(),
+            Permissions::SEND_MESSAGES,
+            &PermissionOverwriteType::Member(UserId(2)),
+        );
+        let actual = builder.request().expect("failed to create request");
+
+        let body = crate::json_to_vec(&UpdateChannelPermissionConfiguredFields {
+            allow: Permissions::empty(),
+            deny: Permissions::SEND_MESSAGES,
+            kind: PermissionOverwriteTargetType::Member,
+        })
+        .expect("failed to serialize payload");
+        let route = Route::UpdatePermissionOverwrite {
+            channel_id: 1,
+            target_id: 2,
+        };
+        let expected = Request::from((body, route));
+
+        assert_eq!(expected.body, actual.body);
+        assert_eq!(expected.path, actual.path);
+    }
+}