+++ model/src/gateway/payload/identify.rs
@@ -19,7 +19,7 @@ impl Identify {
 
 #[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq, Serialize)]
 pub struct IdentifyInfo {
-    pub compression: bool,
+    pub compress: bool,
     pub intents: Intents,
     pub large_threshold: u64,
     pub presence: Option<UpdateStatusInfo>,