+++ standby/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Changelog for `twilight-standby`.
 
+## [0.2.1] - 2020-11-29
+
+### Misc.
+
+Use the renamed
+`twilight_model::gateway::payload::identify::IdentityInfo::compress`
+model field ([#624] - [@chamburr]).
+
 ## [0.2.0] - 2020-10-30
 
 This major version bump of Standby is done to match all of the other crates in
@@ -22,10 +30,13 @@ the ecosystem receiving a major version bump. There are no changes.
 
 Initial release.
 
+[@chamburr]: https://github.com/chamburr
 [@nickelc]: https://github.com/nickelc
 
+[#624]: https://github.com/twilight-rs/twilight/pull/624
 [#523]: https://github.com/twilight-rs/twilight/pull/523
 
+[0.2.1]: https://github.com/twilight-rs/twilight/releases/tag/standby-v0.2.1
 [0.2.0]: https://github.com/twilight-rs/twilight/releases/tag/standby-v0.2.0
 [0.2.0-beta.0]: https://github.com/twilight-rs/twilight/releases/tag/standby-v0.2.0-beta.0
 [0.1.1]: https://github.com/twilight-rs/twilight/releases/tag/standby-v0.1.1