+++ twilight/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Changelog for `twilight`.
 
+## [0.2.2] - 2020-11-24
+
+### Enhancements
+
+Add links to crates when mentioned in the README ([#610] - [@vivian]).
+
 ## [0.2.1] - 2020-11-02
 
 Update the installation instructions to note version 0.2 instead of version
@@ -23,9 +29,11 @@ Initial release.
 
 [@vivian]: https://github.com/vivian
 
+[#610]: https://github.com/twilight-rs/twilight/pull/610
 [#588]: https://github.com/twilight-rs/twilight/pull/588
 [#498]: https://github.com/twilight-rs/twilight/pull/498
 
+[0.2.2]: https://github.com/twilight-rs/twilight/releases/tag/twilight-v0.2.2
 [0.2.1]: https://github.com/twilight-rs/twilight/releases/tag/twilight-v0.2.1
 [0.2.0]: https://github.com/twilight-rs/twilight/releases/tag/twilight-v0.2.0
 [0.1.1]: https://github.com/twilight-rs/twilight/releases/tag/twilight-v0.1.1